Replace Uint128 with Decimal for Rate
