+++ contracts/andromeda_rates/src/contract.rs
@@ -215,13 +215,14 @@ mod tests {
     use super::*;
     use crate::contract::{execute, instantiate, query};
     use andromeda_protocol::{
-        modules::{ADORate, Rate},
-        rates::{InstantiateMsg, PaymentsResponse, QueryMsg, RateInfo},
+        rates::{ADORate, InstantiateMsg, PaymentsResponse, QueryMsg, Rate, RateInfo},
         testing::mock_querier::{mock_dependencies_custom, MOCK_PRIMITIVE_CONTRACT},
     };
     use common::{ado_base::recipient::Recipient, encode_binary};
     use cosmwasm_std::testing::{mock_dependencies, mock_env, mock_info};
-    use cosmwasm_std::{coin, coins, from_binary, BankMsg, Coin, CosmosMsg, Uint128, WasmMsg};
+    use cosmwasm_std::{
+        coin, coins, from_binary, BankMsg, Coin, CosmosMsg, Decimal, Uint128, WasmMsg,
+    };
     use cw20::Cw20ExecuteMsg;
 
     #[test]
@@ -232,7 +233,7 @@ mod tests {
         let info = mock_info(owner, &[]);
         let rates = vec![
             RateInfo {
-                rate: Rate::Percent(10u128.into()),
+                rate: Rate::from(Decimal::percent(10)),
                 is_additive: true,
                 description: Some("desc1".to_string()),
                 receivers: vec![Recipient::Addr("".into())],
@@ -274,7 +275,7 @@ mod tests {
         let info = mock_info(owner, &[]);
         let rates = vec![
             RateInfo {
-                rate: Rate::Percent(10u128.into()),
+                rate: Rate::from(Decimal::percent(10)),
                 is_additive: true,
                 description: Some("desc1".to_string()),
                 receivers: vec![Recipient::Addr("".into())],
@@ -319,7 +320,7 @@ mod tests {
                 receivers: vec![Recipient::Addr("1".into())],
             },
             RateInfo {
-                rate: Rate::Percent(10u128.into()),
+                rate: Rate::from(Decimal::percent(10)),
                 is_additive: false,
                 description: Some("desc1".to_string()),
                 receivers: vec![Recipient::Addr("2".into())],
@@ -405,7 +406,7 @@ mod tests {
                 receivers: vec![Recipient::Addr("1".into())],
             },
             RateInfo {
-                rate: Rate::Percent(10u128.into()),
+                rate: Rate::from(Decimal::percent(10)),
                 is_additive: false,
                 description: Some("desc1".to_string()),
                 receivers: vec![Recipient::Addr("2".into())],
