+++ packages/andromeda_protocol/src/modules/mod.rs
@@ -3,96 +3,19 @@ pub mod common;
 pub mod hooks;
 pub mod receipt;
 
-use ::common::{
-    ado_base::query_get,
-    encode_binary,
-    error::ContractError,
-    primitive::{GetValueResponse, Primitive},
-    require,
-};
+use ::common::error::ContractError;
 
 use crate::modules::{
     address_list::AddressListModule,
     hooks::{HookResponse, MessageHooks},
     receipt::ReceiptModule,
 };
-use cosmwasm_std::{Coin, DepsMut, Env, MessageInfo, QuerierWrapper, StdResult, Storage, Uint128};
+use cosmwasm_std::{DepsMut, Env, MessageInfo, QuerierWrapper, StdResult, Storage};
 use cw_storage_plus::Item;
 use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
 
 pub const MODULES: Item<Modules> = Item::new("modules");
-
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
-#[serde(rename_all = "snake_case")]
-pub struct ADORate {
-    /// The address of the primitive contract.
-    pub address: String,
-    /// The key of the primitive in the primitive contract.
-    pub key: Option<String>,
-}
-
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
-#[serde(rename_all = "snake_case")]
-/// An enum used to define various types of fees
-pub enum Rate {
-    /// A flat rate fee
-    Flat(Coin),
-    /// A percentage fee (integer)
-    Percent(Uint128),
-    External(ADORate),
-}
-
-impl Rate {
-    /// Validates that a given rate is non-zero. It is expected that the Rate is not an
-    /// External Rate.
-    pub fn is_non_zero(&self) -> Result<bool, ContractError> {
-        match self {
-            Rate::Flat(coin) => Ok(coin.amount > Uint128::zero()),
-            Rate::Percent(rate) => Ok(rate > &Uint128::zero()),
-            Rate::External(_) => Err(ContractError::UnexpectedExternalRate {}),
-        }
-    }
-
-    /// Validates `self` and returns an "unwrapped" version of itself wherein if it is an External
-    /// Rate, the actual rate value is retrieved from the Primitive Contract.
-    pub fn validate(&self, querier: &QuerierWrapper) -> Result<Rate, ContractError> {
-        let rate = self.clone().get_rate(querier)?;
-        require(rate.is_non_zero()?, ContractError::InvalidRate {})?;
-
-        if let Rate::Percent(rate) = rate {
-            require(
-                rate <= Uint128::from(100u128),
-                ContractError::InvalidRate {},
-            )?;
-        }
-
-        Ok(rate)
-    }
-
-    /// If `self` is Flat or Percent it returns itself. Otherwise it queries the primitive contract
-    /// and retrieves the actual Flat or Percent rate.
-    fn get_rate(self, querier: &QuerierWrapper) -> Result<Rate, ContractError> {
-        match self {
-            Rate::Flat(_) => Ok(self),
-            Rate::Percent(_) => Ok(self),
-            Rate::External(ado_rate) => {
-                let response: GetValueResponse = query_get(
-                    Some(encode_binary(&ado_rate.key)?),
-                    ado_rate.address,
-                    querier,
-                )?;
-                match response.value {
-                    Primitive::Coin(coin) => Ok(Rate::Flat(coin)),
-                    Primitive::Uint128(value) => Ok(Rate::Percent(value)),
-                    _ => Err(ContractError::ParsingError {
-                        err: "Stored rate is not a coin or Uint128".to_string(),
-                    }),
-                }
-            }
-        }
-    }
-}
 
 #[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
 #[serde(rename_all = "snake_case")]
@@ -265,33 +188,4 @@ pub fn generate_instantiate_msgs(
     }
 
     Ok(resp)
-}
-
-#[cfg(test)]
-mod tests {
-    use crate::testing::mock_querier::{mock_dependencies_custom, MOCK_PRIMITIVE_CONTRACT};
-    use cosmwasm_std::coin;
-
-    use super::*;
-
-    #[test]
-    fn test_validate_external_rate() {
-        let mut deps = mock_dependencies_custom(&[]);
-
-        let rate = Rate::External(ADORate {
-            address: MOCK_PRIMITIVE_CONTRACT.to_string(),
-            key: Some("percent".to_string()),
-        });
-        let validated_rate = rate.validate(&deps.as_mut().querier).unwrap();
-        let expected_rate = Rate::Percent(1u128.into());
-        assert_eq!(expected_rate, validated_rate);
-
-        let rate = Rate::External(ADORate {
-            address: MOCK_PRIMITIVE_CONTRACT.to_string(),
-            key: Some("flat".to_string()),
-        });
-        let validated_rate = rate.validate(&deps.as_mut().querier).unwrap();
-        let expected_rate = Rate::Flat(coin(1u128, "uusd"));
-        assert_eq!(expected_rate, validated_rate);
-    }
 }
